Refactor configuration details
