+++ pages/api/configuration.ts
@@ -26,7 +26,7 @@ const handler: NextApiHandler = async (request, response) => {
   } catch (e: unknown) {
     const error = e as MiddlewareError;
 
-    console.error(error);
+    console.error(error); // For deployment debug purpose
     response.status(error.statusCode).json({
       success: false,
       error,
@@ -60,7 +60,6 @@ const handler: NextApiHandler = async (request, response) => {
           )
           .toPromise()
       ).data?.app?.privateMetafields;
-      console.log("privateMetadata", privateMetadata);
       if (privateMetadata) {
         response.json({
           success: true,
