+++ pages/api/webhooks/order-created.ts
@@ -22,7 +22,7 @@ const handler: NextApiHandler = async (request, response) => {
   } catch (e: unknown) {
     const error = e as MiddlewareError;
 
-    console.error(error);
+    console.error(error); // For deployment debug purpose
     response
       .status(error.statusCode)
       .json({ success: false, message: error.message });
